Refactored web modeler interactions to separate class
